refactor(tauri): PlainTextコンポーネントをsectionディレクトリに抽出
TextFormElement.tsxのPlainTextとDatasetTextFormElement.tsxのDatasetPlainTextの
重複JSXをPlainText.tsxに統合。onValueChangeコールバックで差分を吸収する。
