+++ tauri/src/app/form/section/DatasetTextFormElement.tsx
@@ -1,5 +1,3 @@
-import { useState } from "react";
-import { ControllTextBox, InputLabel } from "../../../components/element/Input";
 import {
 	useDatasetSrcInfo,
 	useSetDatasetSrcInfo,
@@ -9,7 +7,7 @@ import { isSqlRelatedType } from "../../../model/QueryDatasource";
 import DatasetFileText from "./DatasetFileText";
 import DatasetSettingText from "./DatasetSettingText";
 import type { Prop } from "./FormElementProp";
-import { getId, getName } from "./FormElementProp";
+import PlainText from "./PlainText";
 import SqlSrcText from "./SqlSrcText";
 import TemplateText from "./TemplateText";
 import XlsxSchemaText from "./XlsxSchemaText";
@@ -37,15 +35,10 @@ export default function DatasetText(prop: Prop) {
 }
 
 function DatasetPlainText({ prefix, element, hidden }: Prop) {
-	const [path, setPath] = useState(element.value);
 	const datasetSrcInfo = useDatasetSrcInfo();
 	const setDatasetSrcInfo = useSetDatasetSrcInfo();
-	const id = getId(prefix, element.name);
-	const fieldName = getName(prefix, element.name);
 
-	const handleChange = (ev: React.ChangeEvent<HTMLInputElement>) => {
-		const newValue = ev.target.value;
-		setPath(newValue);
+	const handleValueChange = (newValue: string) => {
 		if (datasetSrcInfo && element.name in datasetSrcInfo) {
 			setDatasetSrcInfo({
 				...datasetSrcInfo,
@@ -55,25 +48,11 @@ function DatasetPlainText({ prefix, element, hidden }: Prop) {
 	};
 
 	return (
-		<div>
-			<InputLabel
-				text={fieldName}
-				id={id}
-				required={element.attribute.required}
-				hidden={hidden}
-			/>
-			<div className="flex">
-				<div className="flex-1 mr-36">
-					<ControllTextBox
-						name={fieldName}
-						id={id}
-						hidden={hidden}
-						required={element.attribute.required}
-						value={path}
-						handleChange={handleChange}
-					/>
-				</div>
-			</div>
-		</div>
+		<PlainText
+			prefix={prefix}
+			element={element}
+			hidden={hidden}
+			onValueChange={handleValueChange}
+		/>
 	);
 }
